결제하기, 정산하기 전까지 커밋

+++ src/main/java/com/metalancer/backend/request/service/RequestServiceImpl.java
@@ -2,6 +2,7 @@ package com.metalancer.backend.request.service;
 
 import com.metalancer.backend.common.config.security.PrincipalDetails;
 import com.metalancer.backend.common.constants.DataStatus;
+import com.metalancer.backend.common.constants.ErrorCode;
 import com.metalancer.backend.common.exception.BaseException;
 import com.metalancer.backend.external.aws.s3.S3Service;
 import com.metalancer.backend.request.domain.ProductsRequest;
@@ -10,13 +11,14 @@ import com.metalancer.backend.request.dto.ProductsRequestDTO.File;
 import com.metalancer.backend.request.dto.ProductsRequestDTO.Update;
 import com.metalancer.backend.request.entity.ProductsRequestEntity;
 import com.metalancer.backend.request.repository.ProductsRequestRepository;
-import java.util.List;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+
+import java.util.List;
 
 @Service
 @Slf4j
@@ -29,7 +31,7 @@ public class RequestServiceImpl implements RequestService {
 
     @Override
     public Page<ProductsRequest> getProductsRequestList(List<String> requestTypeOptions,
-        Pageable adjustedPageable) {
+                                                        Pageable adjustedPageable) {
         return productsRequestRepository.findAll(requestTypeOptions, adjustedPageable);
     }
 
@@ -41,7 +43,10 @@ public class RequestServiceImpl implements RequestService {
     @Override
     public ProductsRequest updateRequest(PrincipalDetails user, Update dto, Long requestId) {
         ProductsRequestEntity productsRequestEntity = productsRequestRepository.findEntity(
-            requestId);
+                requestId);
+        if (productsRequestEntity.getWriter() != user.getUser()) {
+            throw new BaseException(ErrorCode.IS_NOT_WRITER);
+        }
         return productsRequestRepository.update(user.getUser(), dto, productsRequestEntity);
     }
 
@@ -53,21 +58,24 @@ public class RequestServiceImpl implements RequestService {
     @Override
     public boolean deleteRequest(PrincipalDetails user, Long requestId) {
         ProductsRequestEntity productsRequestEntity = productsRequestRepository.findEntity(
-            requestId);
+                requestId);
+        if (productsRequestEntity.getWriter() != user.getUser()) {
+            throw new BaseException(ErrorCode.IS_NOT_WRITER);
+        }
         productsRequestEntity.deleteRequest();
         return productsRequestEntity.getStatus().equals(DataStatus.DELETED);
     }
 
     @Override
     public String getUploadRequestFilePreSignedUrl(PrincipalDetails user, Long requestId,
-        String fileName) {
+                                                   String fileName) {
         return s3Service.getRequestFilePresignedUrl(requestId, fileName);
     }
 
     @Override
     public boolean updateRequestFile(PrincipalDetails user, Long requestId, File dto) {
         ProductsRequestEntity productsRequestEntity = productsRequestRepository.findEntity(
-            requestId);
+                requestId);
         productsRequestEntity.setFile(dto.getFileUrl(), dto.getFileName());
         return true;
     }
